Company_Document_QnA/requirements.txt
langchain 
langchain-community 
langchain-openai 
pypdf 
tiktoken 
python-dotenv
streamlit
chromadb
langchain-chroma
pyngrok
pdfplumber
doc2text
unstructured
docx
symspellpy